Moving readHandshake and related calls from ServerHandshakeProcessor to new ServerSideHandshakeFactory
